fix(gl): drop now-redundant int64_t cast on memTextures
clang-tidy readability-redundant-casting: the cast was written when
RenderingStats::memTextures was int; it is int64_t now, so the cast is a
no-op. The other static_cast<int64_t> calls here convert size_t and stay.

+++ src/mbgl/gl/resource_pool.cpp
@@ -204,7 +204,7 @@ TextureID Texture2DPool::allocateGLMemory(const Texture2DDesc& desc) {
 }
 
 void Texture2DPool::checkStatsConsistency(const char* where) {
-    if (statsDivergenceReported || static_cast<int64_t>(context->renderingStats().memTextures) == trackedStorage) {
+    if (statsDivergenceReported || context->renderingStats().memTextures == trackedStorage) {
         return;
     }
     statsDivergenceReported = true;
